# openapi.yaml
openapi: 3.0.3
info:
  title: FitReg API
  description: |
    REST API for FitReg, a fitness coaching platform where coaches assign and manage
    workouts for their students.

    **Authentication:** All endpoints except `/health` and `POST /api/auth/google`
    require a JWT Bearer token obtained via Google OAuth.
  version: "1.0.0"
servers:
  - url: http://localhost:8080
    description: Local development
security:
  - bearerAuth: []
...
paths:
...
  /api/coach/weekly-templates:
    get:
      tags: [coach-weekly-templates]
      summary: List weekly templates
      responses:
        "200":
          description: Weekly templates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WeeklyTemplate"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      tags: [coach-weekly-templates]
      summary: Create weekly template
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                description:
                  type: string
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WeeklyTemplate"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Missing or invalid JWT
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
...
    WorkoutSegment:
      type: object
      properties:
        id:
          type: integer
          format: int64
        order_index:
          type: integer
        segment_type:
          type: string
          enum: [simple, interval]
        repetitions:
          type: integer
        value:
          type: number
          format: float
        unit:
          type: string
          example: km
        intensity:
          type: string
          example: easy
        work_value:
          type: number
          format: float
        work_unit:
          type: string
        work_intensity:
          type: string
        rest_value:
          type: number
          format: float
        rest_unit:
          type: string
        rest_intensity:
          type: string
...
    WeeklyTemplate:
      type: object
      properties:
        id:
          type: integer
          format: int64
        coach_id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        day_count:
          type: integer
          description: Number of days with workouts defined
        days:
          type: array
          items:
            $ref: "#/components/schemas/WeeklyTemplateDay"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    WeeklyTemplateDay:
      type: object
      properties:
        id:
          type: integer
          format: int64
        weekly_template_id:
          type: integer
          format: int64
        day_of_week:
          type: integer
          minimum: 0
          maximum: 6
          description: 0=Monday, 1=Tuesday, ..., 6=Sunday
        title:
          type: string
        description:
          type: string
        type:
          type: string
        distance_km:
          type: number
          format: float
        duration_seconds:
          type: integer
        notes:
          type: string
        from_template_id:
          type: integer
          format: int64
          nullable: true
        segments:
          type: array
          items:
            $ref: "#/components/schemas/WorkoutSegment"
...